Fix package name
This module has not shipped yet, so it's safe to change.
