+++ retrofit-converters/jaxb3/src/main/java/retrofit2/converter/jaxb3/JaxbResponseConverter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package retrofit2.converter.jaxb;
+package retrofit2.converter.jaxb3;
 
 import jakarta.xml.bind.JAXBContext;
 import jakarta.xml.bind.JAXBException;
