+++ retrofit-converters/jaxb3/src/main/java/retrofit2/converter/jaxb3/package-info.java
@@ -1,2 +1,2 @@
 @retrofit2.internal.EverythingIsNonNull
-package retrofit2.converter.jaxb;
+package retrofit2.converter.jaxb3;
