+++ retrofit-converters/jaxb3/src/test/java/retrofit2/converter/jaxb3/PhoneNumber.java
@@ -13,41 +13,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package retrofit2.converter.jaxb;
+package retrofit2.converter.jaxb3;
 
+import jakarta.xml.bind.annotation.XmlAttribute;
 import jakarta.xml.bind.annotation.XmlElement;
-import jakarta.xml.bind.annotation.XmlRootElement;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
+import javax.annotation.Nullable;
 
-@XmlRootElement(name = "contact")
-final class Contact {
+final class PhoneNumber {
   @XmlElement(required = true)
-  public final String name;
+  public final String number;
 
-  @XmlElement(name = "phone_number")
-  public final List<PhoneNumber> phone_numbers;
+  @XmlAttribute public final Type type;
 
   @SuppressWarnings("unused") // Used by JAXB.
-  private Contact() {
-    this("", new ArrayList<PhoneNumber>());
+  private PhoneNumber() {
+    this("", Type.OTHER);
   }
 
-  public Contact(String name, List<PhoneNumber> phoneNumbers) {
-    this.name = name;
-    this.phone_numbers = phoneNumbers;
+  PhoneNumber(String number, @Nullable Type type) {
+    this.number = number;
+    this.type = type;
   }
 
   @Override
   public boolean equals(Object o) {
-    return o instanceof Contact
-        && ((Contact) o).name.equals(name)
-        && ((Contact) o).phone_numbers.equals(phone_numbers);
+    return o instanceof PhoneNumber
+        && ((PhoneNumber) o).number.equals(number)
+        && ((PhoneNumber) o).type.equals(type);
   }
 
   @Override
   public int hashCode() {
-    return Arrays.asList(name, phone_numbers).hashCode();
+    return Arrays.asList(number, type).hashCode();
   }
 }
